replaces tuple with separate class
Tuples are not that great...

+++ CCExtractorTester/Analyzers/Processor.cs
@@ -33,7 +33,7 @@ namespace CCExtractorTester.Analyzers
 
             string commandToPass = String.Format("{0} --no_progress_bar",test.Command);
             string inputFile = Path.Combine(Config.SampleFolder, test.InputFile);
-            string firstOutputFile = Path.Combine(Config.TemporaryFolder, test.CompareFiles[0].Item1);
+            string firstOutputFile = Path.Combine(Config.TemporaryFolder, test.CompareFiles[0].CorrectFile);
 
             FileInfo firstOutputFileFI = new FileInfo(firstOutputFile);
 
@@ -228,24 +228,27 @@ namespace CCExtractorTester.Analyzers
             // Add generated output file(s), if any, to the result
             rd.ResultFiles = new Dictionary<string, string>();
             // Check for each expected output file if there's a generated one
-            foreach (Tuple<string,string,bool> expectedFile in test.CompareFiles)
+            foreach (CompareFile compareFile in test.CompareFiles)
             {
                 // TODO: fix!
-                FileInfo fileLocation = new FileInfo(Path.Combine(Config.TemporaryFolder, expectedFile.Item1));
-                string moveLocation = Path.Combine(storeDirectory, fileLocation.Name);
-                if (fileLocation.Exists)
+                if (!compareFile.IgnoreOutput)
                 {
-                    rd.ResultFiles.Add(expectedFile.Item1, moveLocation);
-                    // Move file
-                    if (File.Exists(moveLocation))
+                    FileInfo fileLocation = new FileInfo(Path.Combine(Config.TemporaryFolder, compareFile.ExpectedFile));
+                    string moveLocation = Path.Combine(storeDirectory, (new FileInfo(Path.Combine(Config.TemporaryFolder,compareFile.CorrectFile)).Name));
+                    if (fileLocation.Exists)
                     {
-                        File.Delete(moveLocation);
+                        rd.ResultFiles.Add(compareFile.CorrectFile, moveLocation);
+                        // Move file
+                        if (File.Exists(moveLocation))
+                        {
+                            File.Delete(moveLocation);
+                        }
+                        fileLocation.MoveTo(moveLocation);
                     }
-                    fileLocation.MoveTo(moveLocation);
-                }
-                else
-                {
-                    rd.ResultFiles.Add(expectedFile.Item1, "");
+                    else
+                    {
+                        rd.ResultFiles.Add(compareFile.CorrectFile, "");
+                    }
                 }
             }
             // Return the results
